requirements.txt
asgiref==3.5.2
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
coreapi==2.3.3
coreschema==0.0.4
cryptography==38.0.4
defusedxml==0.7.1
Django==4.1.3
django-allauth==0.51.0
django-ckeditor==6.5.1
django-debug-toolbar==3.7.0
django-js-asset==2.0.0
djangorestframework==3.14.0
drf-yasg==1.21.4
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.1
oauthlib==3.2.2
packaging==21.3
Pillow==9.3.0
pycparser==2.21
PyJWT==2.6.0
pyparsing==3.0.9
python3-openid==3.2.0
pytz==2022.6
requests==2.28.1
requests-oauthlib==1.3.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
sqlparse==0.4.3
uritemplate==4.1.1
urllib3==1.26.13
